+++ app/helpers/users/payments_helper.rb
@@ -1,0 +1,2 @@
+module Users::PaymentsHelper
+end
