+++ app/views/users/payments/index.html.erb
@@ -1,0 +1,35 @@
+
+
+
+	<div>
+		<p>毎月¥100 で　100Point </p>
+		<%= form_for(@point,url: {controler: "users/payments", action: "monthly_subscription" }, html: { class: "form-horizontal"} ) do |f| %>
+			<%= f.hidden_field :point, value: 100 %><br>
+  			<%= f.submit '¥100プランにする' %>
+  	 	<% end %>
+	</div>
+
+	<div>
+		<p>毎月¥200 で　240Point </p>
+		<%= form_for(@point,url: {controler: "users/payments", action: "monthly_subscription" }, html: { class: "form-horizontal"} ) do |f| %>
+			<%= f.hidden_field :point, value: 240 %><br>
+  			<%= f.submit '¥200プランにする' %>
+  	 	<% end %>
+	</div>
+
+
+	<div>
+		<p>毎月¥300 で　390Point </p>
+		<%= form_for(@point,url: {controler: "users/payments", action: "monthly_subscription" }, html: { class: "form-horizontal"} ) do |f| %>
+			<%= f.hidden_field :point, value: 390 %><br>
+  			<%= f.submit '¥300プランにする' %>
+  	 	<% end %>
+	</div>
+
+	<div>
+		<p>毎月¥500 で　750Point </p>
+		<%= form_for(@point,url: {controler: "users/payments", action: "monthly_subscription" }, html: { class: "form-horizontal"} ) do |f| %>
+			<%= f.hidden_field :point, value: 750 %><br>
+  			<%= f.submit '¥500プランにする' %>
+  	 	<% end %>
+	</div>
