+++ app/views/users/payments/new.html.erb
@@ -1,0 +1,4 @@
+new
+<%= form_tag(action: :registration_payjp, method: :post) do %>
+   <script type="text/javascript" src="https://checkout.pay.jp" class="payjp-button" data-key="pk_test_932adb5c9baabfd8d99fadb3"></script>
+<% end %>
