+++ app/views/users/users/pay.html.erb
@@ -1,1 +1,2 @@
 ok
+<%= button_to '100円', {controller: 'users/users', action: 'monthly_subscription'}, {method: :post, class: "btn btn-default"} %>
