+++ config/routes.rb
@@ -7,8 +7,9 @@ Rails.application.routes.draw do
 	get 'top' => 'top#top'
 	root :to => 'oauth_test#index'
 	get '/notes/hashtag/:tag_name', to: "users/notes#hashtag"#リンクをつけたハッシュタグへのrouting
-  post 'users/points' => 'users/users#registration_payjp'
-  post 'users/points' => 'users/users#pay'
+  post '/payments/new' => 'users/payments#registration_payjp'
+  post 'users/points' => 'users/payments#pay'
+  post '/payments' => 'users/payments#monthly_subscription'
 
 	#github認証のため以下記載
 	devise_for :users, controllers: {
@@ -19,7 +20,7 @@ Rails.application.routes.draw do
 	# For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
   	scope module: :users do
-  		resources :bookmarks, :hashtags, :monthly_fees, :my_notes, :note_hashtags, :points, :users
+  		resources :bookmarks, :hashtags, :monthly_fees, :my_notes, :note_hashtags, :points, :users, :payments
   	end
 
   	scope module: :users do
